requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2025.10.5
click==8.3.0
fastapi==0.115.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.1.1
redis==5.0.8
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.15.0
uvicorn==0.30.6
